pyproject.toml
[project]
name = "study-site"
version = "0.1.0"
description = ""
authors = [
    {name = "Dmitrii Shuliak",email = "x350sh@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "pytest (>=8.4.0,<9.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "django-nested-admin (>=4.1.1,<5.0.0)",
    "django-select2 (>=8.4.1,<9.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "django-extra-views (>=0.16.0,<0.17.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
